requirements.txt
# Core FastAPI dependencies
fastapi==0.115.5
uvicorn[standard]==0.30.6
python-multipart==0.0.9
slowapi==0.1.9

# Async SQLAlchemy + SQLite
aiosqlite>=0.20.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.1

# Werkzeug still needed for secure_filename
Werkzeug==3.1.3

# Database & Auth
PyJWT==2.9.0
bcrypt==4.2.0

# PDF processing
pdfplumber==0.11.7
PyMuPDF==1.26.3
pdfminer.six==20250506

# Document processing
python-docx==1.1.2
pytesseract==0.3.13

# AI — dual-provider (set AI_PROVIDER=gemini or openai in .env)
google-generativeai>=0.3.0
openai==1.95.1

# Environment and configuration
python-dotenv==1.1.1

# Utilities
requests==2.32.4
Pillow==11.3.0

# Vector store (numpy-based, stored in SQLite — replaces chromadb)
numpy>=1.26.0

# MCP (Model Context Protocol)
mcp>=1.0.0

# Async task queue
celery[redis]==5.4.0
redis==5.2.1

# Rate limiting (distributed via Redis)
# Flask-Limiter replaced by slowapi (added above)

# Structured logging
structlog==25.1.0

# AWS S3 (optional file storage)
boto3==1.38.0

# Web search + scraping (Explore Hub)
ddgs>=0.1.0
trafilatura>=1.12.0

# Production server (use: gunicorn -w 4 -k uvicorn.workers.UvicornWorker main:app)
gunicorn==23.0.0

# Socket.IO for real-time indexing progress
python-socketio==5.11.4

